TbMapLocation becomes a short

+++ src/map_locations.h
@@ -29,7 +29,7 @@ extern "C" {
 /******************************************************************************/
 
 
-typedef unsigned long TbMapLocation;
+typedef unsigned short TbMapLocation;
 
 enum MapLocationTypes {
     MLoc_NONE = 0,
